Adding hook usePlaceOrder
